requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.2
ftfy==6.3.1
grpclib==0.4.7
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
modal==0.73.138
mpmath==1.3.0
multidict==6.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.2.0
propcache==0.3.1
protobuf==5.29.4
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
setuptools==78.1.0
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
synchronicity==0.9.11
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typer==0.15.2
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
watchfiles==1.0.4
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3
